+++ floor2drift/test/flutter_test/test_databases/entities/test_entity.dart
@@ -1,4 +1,9 @@
 import 'package:floor/floor.dart';
+// DO NOT REMOVE. Needed to test if Table class is hidden
+// ignore: unused_import needed to test if table is being hidden
+import "package:flutter/cupertino.dart";
+// ignore: unused_import needed to test if table is being hidden
+import 'package:flutter/material.dart';
 
 import '../support_classes/enums.dart';
 import 'test_base_entity_two.dart';