+++ floor2drift/CHANGELOG.md
@@ -2,7 +2,7 @@
 - entity wasn't imported if the dao only had methods inherited from a base dao
 - files written use the current platforms line terminator instead of always '\n'
 - fixed problem where git indicated files changed even if the content was the same.
-- hide material/cupertino table import in (base-)dao classes to fix  name collision of `Table` class
+- hide material/cupertino table import in (base-)entity classes to fix  name collision of `Table` class
 
 ## 0.1.7
 - rolled back dart sdk dependency to ^3.6.0